Solve select n+1 issue

+++ src/main/resources/schema.sql
@@ -26,13 +26,13 @@ CONSTRAINT fk_response_survey FOREIGN KEY (survey_id) REFERENCES survey(id)
 );
 
 CREATE TABLE rating (
-id INT AUTO_INCREMENT PRIMARY KEY,
+id BIGINT AUTO_INCREMENT PRIMARY KEY,
 label VARCHAR(50) NOT NULL
 );
 
 CREATE TABLE answer (
 id BIGINT AUTO_INCREMENT PRIMARY KEY,
-rating_id INT NOT NULL,
+rating_id BIGINT NOT NULL,
 question_id BIGINT,
 response_id BIGINT,
 CONSTRAINT fk_answer_rating FOREIGN KEY (rating_id) REFERENCES rating(id),
